+++ capacity-checker-impl/src/main/java/com/dwarfeng/capacitychecker/impl/handler/CheckerHandlerImpl.java
@@ -9,7 +9,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
-import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 
@@ -21,7 +21,7 @@ public class CheckerHandlerImpl implements CheckerHandler {
     private final List<CheckerMaker> checkerMakers;
 
     public CheckerHandlerImpl(List<CheckerMaker> checkerMakers) {
-        this.checkerMakers = Objects.isNull(checkerMakers) ? new ArrayList<>() : checkerMakers;
+        this.checkerMakers = Objects.isNull(checkerMakers) ? Collections.emptyList() : checkerMakers;
     }
 
     @Override