+++ capacity-checker-impl/src/main/java/com/dwarfeng/capacitychecker/impl/handler/pusher/LogPusher.java
@@ -1,0 +1,89 @@
+package com.dwarfeng.capacitychecker.impl.handler.pusher;
+
+import com.alibaba.fastjson.JSON;
+import com.dwarfeng.capacitychecker.sdk.bean.entity.FastJsonAlarmInfo;
+import com.dwarfeng.capacitychecker.sdk.bean.entity.FastJsonCheckHistory;
+import com.dwarfeng.capacitychecker.stack.bean.entity.AlarmInfo;
+import com.dwarfeng.capacitychecker.stack.bean.entity.CheckHistory;
+import com.dwarfeng.subgrade.stack.exception.HandlerException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+/**
+ * 将信息输出至日志的推送器。
+ *
+ * @author DwArFeng
+ * @since 1.0.0
+ */
+@Component
+public class LogPusher extends AbstractPusher {
+
+    public static final String PUSHER_TYPE = "log";
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(LogPusher.class);
+
+    private static final String LEVEL_TRACE = "TRACE";
+    private static final String LEVEL_DEBUG = "DEBUG";
+    private static final String LEVEL_INFO = "INFO";
+    private static final String LEVEL_WARN = "WARN";
+    private static final String LEVEL_ERROR = "ERROR";
+
+    @Value("${pusher.log.log_level}")
+    private String logLevel;
+
+    public LogPusher() {
+        super(PUSHER_TYPE);
+    }
+
+    @Override
+    public void alarmInfoUpdated(AlarmInfo alarmInfo) throws HandlerException {
+        String title = "推送报警信息更新事件:";
+        String message = JSON.toJSONString(FastJsonAlarmInfo.of(alarmInfo), true);
+        logData(title, message);
+    }
+
+    @Override
+    public void checkHistoryRecorded(CheckHistory checkHistory) throws HandlerException {
+        String title = "推送检查历史记录事件:";
+        String message = JSON.toJSONString(FastJsonCheckHistory.of(checkHistory), true);
+        logData(title, message);
+    }
+
+    private void logData(String title, String message) throws HandlerException {
+        String logLevel = this.logLevel.toUpperCase();
+        switch (logLevel) {
+            case LEVEL_TRACE:
+                LOGGER.trace(title);
+                LOGGER.trace(message);
+                return;
+            case LEVEL_DEBUG:
+                LOGGER.debug(title);
+                LOGGER.debug(message);
+                return;
+            case LEVEL_INFO:
+                LOGGER.info(title);
+                LOGGER.info(message);
+                return;
+            case LEVEL_WARN:
+                LOGGER.warn(title);
+                LOGGER.warn(message);
+                return;
+            case LEVEL_ERROR:
+                LOGGER.error(title);
+                LOGGER.error(message);
+                return;
+            default:
+                throw new HandlerException("未知的日志等级: " + logLevel);
+        }
+    }
+
+    @Override
+    public String toString() {
+        return "LogPusher{" +
+                "logLevel='" + logLevel + '\'' +
+                ", pusherType='" + pusherType + '\'' +
+                '}';
+    }
+}